Move WorkbenchItems in pieapp/structs/workbench.py
