+++ src/pieapp/structs/workbench.py
@@ -1,0 +1,12 @@
+"""
+Workbench items
+"""
+
+
+class WorkbenchItems:
+    OpenFiles = "openFiles"
+    Clear = "clear"
+    Convert = "convert"
+    Settings = "settings"
+    Spacer = "spacer"
+    Exit = "exit"
